requirements.txt
Flask==1.1.2
Flask_Cors==3.0.9
Flask_MySQLdb==0.2.0
simplejson==3.17.6
Werkzeug==1.0.1
pyyaml==4.2b1
